requirements.txt
fastapi
vllm 
gradio 
einops 
transformers_stream_generator
accelerate
optimum
auto-gptq
